requirements.txt
aiogram>=3.0.0b7
asyncpg
sqlalchemy
python-dotenv